Switched over to uWebSockets entirely
In order to remain partly sane I switched over to uWebSockets from Boost Beast, it had better documentation (although 3/4 of it was for JavaScript), way faster and Express-like syntax, I was struggling a whole ton with Boost Beast and even opened an issue in their GitHub repo as there was nobody around to help. Most of the things you see in this commit are already in place, I just need to fix the search for formdata, add a way to send the ZIP-file to a client and clean up all the comments as well as fix some of the things I stated there.